+++ crates/controller-core/src/api/v1/streaming_targets.rs
@@ -10,26 +10,24 @@ use super::{
 use crate::api::v1::ApiResponse;
 use actix_web::{
     delete, get, patch, post,
-    web::{Data, Json, Path, Query},
+    web::{Data, Json, Path},
 };
 use anyhow::Context;
 use opentalk_database::{Db, DbConnection};
 use opentalk_db_storage::streaming_targets::{
-    RoomStreamingTargetNew, RoomStreamingTargetRecord, UpdateRoomStreamingTarget,
+    NewRoomStreamingTarget, RoomStreamingTargetRecord, UpdateRoomStreamingTarget,
 };
 use opentalk_types::{
     api::v1::{
-        pagination::PagePaginationQuery,
         rooms::streaming_targets::{
             ChangeRoomStreamingTargetRequest, ChangeRoomStreamingTargetResponse,
             GetRoomStreamingTargetResponse, GetRoomStreamingTargetsResponse,
             PostRoomStreamingTargetRequest, PostRoomStreamingTargetResponse,
         },
-        streaming_targets::{RoomAndStreamingTargetId, UpdateStreamingTargetKind},
+        streaming_targets::StreamingTargetIdentifier,
     },
-    common::streaming::{RoomStreamingTarget, StreamingTarget, StreamingTargetKind},
+    common::streaming::{RoomStreamingTarget, StreamingTarget},
     core::RoomId,
-    core::StreamingKind,
 };
 
 /// API Endpoint *GET /rooms/{room_id}/streaming_targets*
@@ -39,20 +37,15 @@ use opentalk_types::{
 pub async fn get_streaming_targets(
     db: Data<Db>,
     room_id: Path<RoomId>,
-    pagination: Query<PagePaginationQuery>,
 ) -> DefaultApiResult<GetRoomStreamingTargetsResponse> {
     let mut conn = db.get_conn().await?;
     let room_id = room_id.into_inner();
-    let PagePaginationQuery { per_page, page } = pagination.into_inner();
 
     let room_streaming_targets = get_room_streaming_targets(&mut conn, room_id).await?;
 
-    let len = room_streaming_targets.len();
-
-    Ok(
-        ApiResponse::new(GetRoomStreamingTargetsResponse(room_streaming_targets))
-            .with_page_pagination(per_page, page, len as i64),
-    )
+    Ok(ApiResponse::new(GetRoomStreamingTargetsResponse(
+        room_streaming_targets,
+    )))
 }
 
 pub(super) async fn get_room_streaming_targets(
@@ -63,27 +56,9 @@ pub(super) async fn get_room_streaming_targets(
 
     let room_streaming_targets = streaming_targets
         .into_iter()
-        .map(|st| {
-            let streaming_endpoint = st
-                .streaming_endpoint
-                .parse()
-                .context("invalid streaming endpoint url entry in db")?;
-            let public_url = st
-                .public_url
-                .parse()
-                .context("invalid public url entry in db")?;
-
-            let room_streaming_target = RoomStreamingTarget {
-                id: st.id,
-                streaming_target: StreamingTarget {
-                    name: st.name,
-                    kind: StreamingTargetKind::Custom {
-                        streaming_endpoint,
-                        streaming_key: st.streaming_key,
-                        public_url,
-                    },
-                },
-            };
+        .map(|room_streaming_target_record| {
+            let room_streaming_target =
+                create_room_streaming_target_from_record(room_streaming_target_record)?;
 
             Ok(room_streaming_target)
         })
@@ -119,19 +94,13 @@ pub(super) async fn insert_room_streaming_target(
 ) -> Result<RoomStreamingTarget, ApiError> {
     let streaming_target_clone = streaming_target.clone();
 
-    let streaming_target_record = match streaming_target.kind {
-        StreamingTargetKind::Custom {
-            streaming_endpoint,
-            streaming_key,
-            public_url,
-        } => RoomStreamingTargetNew {
-            room_id,
-            name: streaming_target_clone.name.clone(),
-            kind: StreamingKind::Custom,
-            streaming_endpoint: streaming_endpoint.into(),
-            streaming_key: streaming_key.into(),
-            public_url: public_url.into(),
-        },
+    let streaming_target_record = NewRoomStreamingTarget {
+        service_id: streaming_target_clone.service_id,
+        room_id,
+        name: streaming_target_clone.name.clone(),
+        streaming_url: streaming_target.streaming_url.map(|s| s.into()),
+        streaming_key: streaming_target.streaming_key.map(|s| s.into()),
+        public_url: streaming_target.public_url.map(|s| s.into()),
     }
     .insert(conn)
     .await?;
@@ -150,35 +119,22 @@ pub(super) async fn insert_room_streaming_target(
 #[get("/rooms/{room_id}/streaming_targets/{streaming_target_id}")]
 pub async fn get_streaming_target(
     db: Data<Db>,
-    path_params: Path<RoomAndStreamingTargetId>,
+    path_params: Path<StreamingTargetIdentifier>,
 ) -> DefaultApiResult<GetRoomStreamingTargetResponse> {
     let mut conn = db.get_conn().await?;
-    let RoomAndStreamingTargetId {
+    let StreamingTargetIdentifier {
         room_id,
         streaming_target_id,
     } = path_params.into_inner();
 
-    let room_streaming_target =
+    let room_streaming_target_record =
         RoomStreamingTargetRecord::get(&mut conn, streaming_target_id, room_id).await?;
 
+    let room_streaming_target =
+        create_room_streaming_target_from_record(room_streaming_target_record)?;
+
     Ok(ApiResponse::new(GetRoomStreamingTargetResponse(
-        RoomStreamingTarget {
-            id: room_streaming_target.id,
-            streaming_target: StreamingTarget {
-                name: room_streaming_target.name,
-                kind: StreamingTargetKind::Custom {
-                    streaming_endpoint: room_streaming_target
-                        .streaming_endpoint
-                        .parse()
-                        .context("invalid streaming endpoint url entry in db")?,
-                    streaming_key: room_streaming_target.streaming_key,
-                    public_url: room_streaming_target
-                        .public_url
-                        .parse()
-                        .context("invalid public url entry in db")?,
-                },
-            },
-        },
+        room_streaming_target,
     )))
 }
 
@@ -188,70 +144,38 @@ pub async fn get_streaming_target(
 #[patch("/rooms/{room_id}/streaming_targets/{streaming_target_id}")]
 pub async fn patch_streaming_target(
     db: Data<Db>,
-    path_params: Path<RoomAndStreamingTargetId>,
-    update_streaming_target: Json<ChangeRoomStreamingTargetRequest>,
+    path_params: Path<StreamingTargetIdentifier>,
+    streaming_target: Json<ChangeRoomStreamingTargetRequest>,
 ) -> DefaultApiResult<ChangeRoomStreamingTargetResponse> {
     let mut conn = db.get_conn().await?;
-    let update_streaming_target = update_streaming_target.into_inner().0;
+    let streaming_target = streaming_target.into_inner().0;
 
-    if update_streaming_target.name.is_none() && update_streaming_target.kind.is_none() {
+    if streaming_target.name.is_none()
+        && streaming_target.streaming_url.is_none()
+        && streaming_target.streaming_key.is_none()
+        && streaming_target.public_url.is_none()
+    {
         return Err(ApiError::bad_request());
     }
 
-    let RoomAndStreamingTargetId {
+    let StreamingTargetIdentifier {
         room_id,
         streaming_target_id,
     } = path_params.into_inner();
 
-    let (kind, streaming_endpoint, streaming_key, public_url) = match update_streaming_target.kind {
-        Some(kind) => match kind {
-            UpdateStreamingTargetKind::Custom {
-                streaming_endpoint,
-                streaming_key,
-                public_url,
-            } => (
-                Some(StreamingKind::Custom),
-                streaming_endpoint.map(|s| s.into()),
-                streaming_key.map(|s| s.into()),
-                public_url.map(|s| s.into()),
-            ),
-        },
-        None => (None, None, None, None),
+    let room_streaming_target_record = UpdateRoomStreamingTarget {
+        name: streaming_target.name,
+        streaming_url: streaming_target.streaming_url.map(|s| s.map(|s| s.into())),
+        streaming_key: streaming_target.streaming_key.map(|s| s.map(|s| s.into())),
+        public_url: streaming_target.public_url.map(|s| s.map(|s| s.into())),
     };
 
-    let room_streaming_target_table = UpdateRoomStreamingTarget {
-        name: update_streaming_target.name,
-        kind,
-        streaming_endpoint,
-        streaming_key,
-        public_url,
-    };
-
-    let room_streaming_target_table = room_streaming_target_table
+    let room_streaming_target_record = room_streaming_target_record
         .apply(&mut conn, room_id, streaming_target_id)
         .await?;
 
-    let kind = match room_streaming_target_table.kind {
-        StreamingKind::Custom => StreamingTargetKind::Custom {
-            streaming_endpoint: room_streaming_target_table
-                .streaming_endpoint
-                .parse()
-                .context("invalid streaming endpoint url entry in db")?,
-            streaming_key: room_streaming_target_table.streaming_key,
-            public_url: room_streaming_target_table
-                .public_url
-                .parse()
-                .context("invalid public url entry in db")?,
-        },
-    };
-
-    let room_streaming_target = RoomStreamingTarget {
-        id: room_streaming_target_table.id,
-        streaming_target: StreamingTarget {
-            name: room_streaming_target_table.name,
-            kind,
-        },
-    };
+    let room_streaming_target =
+        create_room_streaming_target_from_record(room_streaming_target_record)?;
 
     Ok(ApiResponse::new(ChangeRoomStreamingTargetResponse(
         room_streaming_target,
@@ -265,11 +189,11 @@ pub async fn patch_streaming_target(
 #[delete("/rooms/{room_id}/streaming_targets/{streaming_target_id}")]
 pub async fn delete_streaming_target(
     db: Data<Db>,
-    path_params: Path<RoomAndStreamingTargetId>,
+    path_params: Path<StreamingTargetIdentifier>,
 ) -> Result<NoContent, ApiError> {
     let mut conn = db.get_conn().await?;
 
-    let RoomAndStreamingTargetId {
+    let StreamingTargetIdentifier {
         room_id,
         streaming_target_id,
     } = path_params.into_inner();
@@ -277,4 +201,38 @@ pub async fn delete_streaming_target(
     RoomStreamingTargetRecord::delete_by_id(&mut conn, room_id, streaming_target_id).await?;
 
     Ok(NoContent)
+}
+
+fn create_room_streaming_target_from_record(
+    room_streaming_target_record: RoomStreamingTargetRecord,
+) -> Result<RoomStreamingTarget, ApiError> {
+    let streaming_url = if let Some(streaming_url) = room_streaming_target_record.streaming_url {
+        let streaming_url = streaming_url
+            .parse()
+            .context("invalid streaming endpoint url entry in db")?;
+        Some(streaming_url)
+    } else {
+        None
+    };
+
+    let public_url = if let Some(public_url) = room_streaming_target_record.public_url {
+        let public_url = public_url
+            .parse()
+            .context("invalid public url entry in db")?;
+        Some(public_url)
+    } else {
+        None
+    };
+
+    let room_streaming_target = RoomStreamingTarget {
+        id: room_streaming_target_record.id,
+        streaming_target: StreamingTarget {
+            service_id: room_streaming_target_record.service_id,
+            name: room_streaming_target_record.name,
+            streaming_url,
+            streaming_key: room_streaming_target_record.streaming_key,
+            public_url,
+        },
+    };
+    Ok(room_streaming_target)
 }