+++ crates/db-storage/src/lib.rs
@@ -25,6 +25,7 @@ pub mod migrations;
 pub mod module_resources;
 pub mod rooms;
 pub mod sip_configs;
+pub mod streaming_services;
 pub mod streaming_targets;
 pub mod tariffs;
 pub mod tenants;