+++ crates/types/src/common/streaming.rs
@@ -6,40 +6,94 @@
 
 use url::Url;
 
-use crate::core::{StreamingKey, StreamingTargetId};
+use crate::core::{StreamingKey, StreamingServiceId, StreamingTargetId};
 
 #[allow(unused_imports)]
 use crate::imports::*;
 
-/// A streaming target kind
+/// A streaming service kind
 #[derive(Clone, Debug, PartialEq, Eq)]
+#[cfg_attr(feature = "clap", derive(clap::ValueEnum))]
 #[cfg_attr(
     feature = "serde",
     derive(Serialize, Deserialize),
     serde(tag = "kind", rename_all = "snake_case")
 )]
-pub enum StreamingTargetKind {
-    /// The "custom" kind
-    Custom {
-        /// The endpoint url of the streaming target
-        streaming_endpoint: Url,
-        /// The streaming key
-        streaming_key: StreamingKey,
-        /// The url from which the stream can be accessed
-        public_url: Url,
-    },
+pub enum StreamingServiceKind {
+    /// The built-in kind
+    Builtin,
+    /// A custom kind
+    Custom,
+    /// A provider kind
+    Provider,
+}
+
+/// A streaming service
+#[derive(Clone, Debug, PartialEq, Eq)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+pub struct StreamingService {
+    /// The id of the streaming service
+    pub id: StreamingServiceId,
+
+    /// The name of the streaming service
+    pub name: String,
+
+    /// The kind of the streaming service
+    #[cfg_attr(feature = "serde", serde(flatten))]
+    pub kind: StreamingServiceKind,
+
+    /// The endpoint url of the streaming service
+    #[cfg_attr(
+        feature = "serde",
+        serde(default, skip_serializing_if = "Option::is_none")
+    )]
+    pub streaming_url: Option<Url>,
+
+    /// The format of the streaming key
+    #[cfg_attr(
+        feature = "serde",
+        serde(default, skip_serializing_if = "Option::is_none")
+    )]
+    pub streaming_key_regex: Option<String>,
+
+    /// The format of the url from which the stream can be accessed
+    #[cfg_attr(
+        feature = "serde",
+        serde(default, skip_serializing_if = "Option::is_none")
+    )]
+    pub public_url_regex: Option<String>,
 }
 
 /// A streaming target
 #[derive(Clone, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct StreamingTarget {
+    /// The id of the streaming service this streaming target belongs to
+    pub service_id: StreamingServiceId,
+
     /// The name of the streaming target
     pub name: String,
 
-    /// The kind of the streaming target
-    #[cfg_attr(feature = "serde", serde(flatten))]
-    pub kind: StreamingTargetKind,
+    /// The endpoint url of the streaming target
+    #[cfg_attr(
+        feature = "serde",
+        serde(default, skip_serializing_if = "Option::is_none")
+    )]
+    pub streaming_url: Option<Url>,
+
+    /// The streaming key
+    #[cfg_attr(
+        feature = "serde",
+        serde(default, skip_serializing_if = "Option::is_none")
+    )]
+    pub streaming_key: Option<StreamingKey>,
+
+    /// The url from which the stream can be accessed
+    #[cfg_attr(
+        feature = "serde",
+        serde(default, skip_serializing_if = "Option::is_none")
+    )]
+    pub public_url: Option<Url>,
 }
 
 /// A streaming target which is specific for a Room