+++ crates/types/src/core/streaming_kind.rs
@@ -10,13 +10,16 @@ use crate::sql_enum;
 sql_enum!(
     feature_gated:
 
-    #[derive(PartialEq, Eq)]
+    #[derive(PartialEq, Eq, strum::Display)]
     #[cfg_attr(feature="serde", derive(Serialize, Deserialize), serde(rename_all = "snake_case"))]
+    #[strum(serialize_all = "snake_case")]
     StreamingKind,
     "streaming_kind",
     StreamingKindType,
     {
+        Builtin = b"builtin",
         Custom = b"custom",
+        Provider = b"provider",
     }
 );