+++ crates/types/src/core/streaming_service_id.rs
@@ -1,0 +1,44 @@
+// SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
+//
+// SPDX-License-Identifier: EUPL-1.2
+
+use derive_more::{AsRef, Display, From, FromStr, Into};
+use uuid::Uuid;
+
+#[allow(unused_imports)]
+use crate::imports::*;
+
+/// ID of a streaming service
+#[derive(
+    AsRef, Display, From, FromStr, Into, Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord,
+)]
+#[cfg_attr(
+    feature = "redis",
+    derive(redis_args::ToRedisArgs, redis_args::FromRedisValue),
+    to_redis_args(fmt),
+    from_redis_value(FromStr)
+)]
+#[cfg_attr(feature="diesel",
+    derive(DieselNewtype, AsExpression, FromSqlRow),
+    diesel(sql_type = diesel::sql_types::Uuid),
+)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+pub struct StreamingServiceId(Uuid);
+
+impl StreamingServiceId {
+    /// Create a ZERO streaming service id, e.g. for testing purposes
+    pub const fn nil() -> Self {
+        Self(Uuid::nil())
+    }
+
+    /// Create a streaming service id from a number, e.g. for testing purposes
+    pub const fn from_u128(id: u128) -> Self {
+        Self(Uuid::from_u128(id))
+    }
+
+    /// Generate a new random streaming service id
+    #[cfg(feature = "rand")]
+    pub fn generate() -> Self {
+        Self::from(Uuid::new_v4())
+    }
+}